DOCUMENTACIÓ

+++ exercicis/exercici12/Program.cs
@@ -28,10 +28,10 @@ class Program
     static void Main(string[] args)
     {
         Console.Write("Entra el primer número: ");
-        int num_1 = Convert.ToInt32(Console.ReadLine());
+        int num_1 = Convert.ToInt16(Console.ReadLine());
 
         Console.Write("Entra el segon número: ");
-        int num_2 = Convert.ToInt32(Console.ReadLine());
+        int num_2 = Convert.ToInt16(Console.ReadLine());
 
         Console.WriteLine($"{num_1} + {num_2} = {num_1 + num_2}");
         Console.WriteLine($"{num_1} - {num_2} = {num_1 - num_2}");
